Removes extra imgui layer

+++ v6/src/imgui-layer/imgui-layer.cpp
@@ -67,8 +67,8 @@ void ImGuiLayer::begin() {
 void ImGuiLayer::end() {
   ImGuiIO &io = ImGui::GetIO();
   Application &app = Application::get();
-  io.DisplaySize =
-      ImVec2(app.getWindow().GetWidth(), app.getWindow().GetHeight());
+  io.DisplaySize = ImVec2((float)app.getWindow().GetWidth(),
+                          (float)app.getWindow().GetHeight());
 
   // Rendering
   ImGui::Render();
